requirements.txt
alabaster==0.7.9
anaconda-client==1.6.0
anaconda-navigator==1.5
anaconda-project==0.4.1
ansimarkup==1.4.0
appdirs==1.4.3
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.24.0
astroid==1.4.9
astropy==1.3
attrs==19.1.0
autopep8==1.4.4
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.5.3
better-exceptions-fork==0.2.1.post6
bitarray==0.8.1
black==19.3b0
blaze==0.10.1
bokeh==0.12.4
boto==2.45.0
Bottleneck==1.2.0
bytedlogger==0.2.2
certifi==2019.6.16
cffi==1.9.1
chardet==3.0.4
chest==0.2.3
Click==7.0
cloudpickle==0.2.2
clyent==1.2.2
colorama==0.3.7
conda==4.7.5
conda-package-handling==1.3.10
configobj==5.0.6
contextlib2==0.5.4
cryptography==2.6.1
cycler==0.10.0
Cython==0.25.2
cytoolz==0.8.2
dask==0.13.0
datashape==0.5.4
decorator==4.4.0
dill==0.2.5
docutils==0.13.1
et-xmlfile==1.0.1
fake-useragent==0.1.10
fastcache==1.0.2
filelock==3.0.12
findspark==1.3.0
Flask==0.12
Flask-Cors==3.0.2
flatbuffers==1.11
funcsigs==1.0.2
future==0.17.1
gevent==1.2.1
greenlet==0.4.11
h5py==2.6.0
HeapDict==1.0.0
idna==2.8
imagesize==0.7.1
influxdb==5.2.2
ipykernel==5.1.0
ipyparallel==6.2.4
ipython==5.1.0
ipython-genutils==0.2.0
ipywidgets==5.2.2
isort==4.2.5
itsdangerous==0.24
jdcal==1.3
jedi==0.14.0
Jinja2==2.9.4
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==5.0.0
jupyter-core==4.5.0
lazy-object-proxy==1.2.2
libarchive-c==2.8
llvmlite==0.15.0
locket==0.2.0
loguru==0.2.5
logzero==1.5.0
lxml==4.1.1
MarkupSafe==0.23
matplotlib==2.0.0
metakernel==0.24.2
mistune==0.7.3
mpmath==0.19
multipledispatch==0.4.9
nbconvert==4.2.0
nbformat==4.2.0
networkx==1.11
nltk==3.2.2
nose==1.3.7
notebook==4.3.1
numba==0.30.1
numexpr==2.6.1
numpy==1.16.4
numpydoc==0.6.0
odo==0.5.0
openpyxl==2.4.1
pandas==0.19.2
parso==0.5.0
partd==0.3.7
pathlib2==2.2.0
patsy==0.4.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
ply==3.9
prompt-toolkit==1.0.9
protobuf==3.9.0
psutil==5.0.1
ptyprocess==0.5.1
py==1.4.32
py4j==0.10.7
pyasn1==0.1.9
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.17
pycrypto==2.6.1
pycurl==7.43.0.2
pyflakes==1.5.0
Pygments==2.4.0
PyHive==0.6.1
pylint==1.6.4
pyOpenSSL==16.2.0
pyparsing==2.1.4
PySocks==1.7.0
pyspark==2.4.3
pytest==3.0.5
python-dateutil==2.8.0
python-telegram-bot==12.0.0b1
pytz==2016.10
PyYAML==3.12
pyzmq==18.0.2
QtAwesome==0.4.3
qtconsole==4.2.1
QtPy==1.2.1
ray==0.7.2
redis==2.10.5
requests==2.21.0
rope-py3k==0.9.4.post1
sasl==0.2.1
schedule==0.6.0
scikit-image==0.12.3
scikit-learn==0.18.1
scipy==0.18.1
seaborn==0.7.1
setproctitle==1.1.10
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
Sphinx==1.5.1
spyder==3.1.2
spylon==0.3.0
SQLAlchemy==1.1.5
statsmodels==0.6.1
sympy==1.0
tables==3.3.0
terminado==0.6
thrift==0.11.0
thrift-sasl==0.3.0
toml==0.10.0
toolz==0.8.2
tornado==6.0.3
tqdm==4.32.1
traitlets==4.3.2
unicodecsv==0.14.1
urllib3==1.24.2
wcwidth==0.1.7
Werkzeug==0.11.15
widgetsnbextension==1.2.6
wrapt==1.10.8
xlrd==1.0.0
XlsxWriter==0.9.6
xlwings==0.10.2
xlwt==1.2.0
yapf==0.27.0
